pyproject.toml
[project]
name = "video-dl"
version = "2.1.6"
description = "A Flet-based GUI for yt-dlp with hardware-accelerated encoding"
requires-python = ">=3.12"
dependencies = [
    "darkdetect",
    "flet>=0.81",
    "flet-desktop>=0.81",
    "quantiphy",
    "tomlkit",
    "tufup>=0.10",
    "yt-dlp-ejs",
    "yt-dlp-enhanced-progress",
]

[project.optional-dependencies]
dev = [
    "mypy",
    "Pillow",
    "pyinstaller",
    "pytest",
    "pytest-cov",
    "ruff",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["core*", "gui*", "i18n*", "runtime*", "updater*", "utils*", "tools*"]

[project.scripts]
video-dl = "main:main"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["core", "gui", "i18n", "runtime", "updater", "utils", "tools", "sys_vars", "videodl_logger"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_return_any = true
check_untyped_defs = true
ignore_missing_imports = true
exclude = ["build", "dist", "venv", "\\.venv"]
disable_error_code = ["no-any-return", "var-annotated"]

[tool.pytest.ini_options]
testpaths = ["tests"]
